requirements.txt
python-dotenv==1.0.0
tavily-python==0.3.3
requests==2.31.0
gradio==4.7.1
aiohttp==3.9.1
openai==1.54.0
httpx==0.27.0
anthropic==0.39.0
